Move code from CommandHandler to Program
